Add e2e tests and expose CLI library
