+++ crates/openwarp-cli/Cargo.toml
@@ -18,6 +18,7 @@ colored = "2"
 open = "5"
 plugin-sdk = { path = "../plugin-sdk" }
 toml = "0.8"
+wasmtime = "19"
 
 [dev-dependencies]
 rstest = "0.18"
