+++ crates/openwarp-e2e/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "openwarp-e2e"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+cucumber = "0.21"
+tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
+async-trait = "0.1"
+openwarp-cli = { path = "../openwarp-cli" }
+anyhow = "1"
+tempfile = "3"
