+++ crates/openwarp-e2e/features/model.feature
@@ -1,0 +1,5 @@
+Feature: Model
+  Scenario: switch
+    Given a command router
+    When I route "/model llama2"
+    Then the route should be Switch "llama2"
