+++ crates/openwarp-e2e/features/safety.feature
@@ -1,0 +1,5 @@
+Feature: Safety
+  Scenario: dangerous command
+    Given a command router
+    When safety scan "rm -rf /"
+    Then scan result is false
