+++ crates/openwarp-e2e/features/smart_mode.feature
@@ -1,0 +1,5 @@
+Feature: Smart Mode
+  Scenario: interactive shell
+    Given a command router
+    When I route "python"
+    Then the route should spawn "python"
