+++ crates/openwarp-e2e/features/translate.feature
@@ -1,0 +1,5 @@
+Feature: Translate
+  Scenario: simple
+    Given a command router
+    When I translate "list files"
+    Then the command should be "echo list files"
